resolve PR threads
