+++ deps.ts
@@ -9,6 +9,7 @@ export { decodeBase64 } from 'https://deno.land/std@0.203.0/encoding/base64.ts'
 export { decode as decodeVarint } from 'https://deno.land/x/varint@v2.0.0/varint.ts'
 export { retry } from 'https://deno.land/std@0.203.0/async/retry.ts';
 
+
 export { ethers } from "https://cdn.skypack.dev/ethers@5.7.2?dts";
 // Deno Bundle does not support npm dependencies, we have to load them via CDN
 export { CarBlockIterator } from 'https://cdn.skypack.dev/@ipld/car@5.3.2/?dts'
