+++ lib/miner-info.js
@@ -30,9 +30,9 @@ async function getChainHead ({ maxAttempts = 5 } = {}) {
  * @returns {Promise<string>} Miner's PeerId, e.g. `12D3KooWMsPmAA65yHAHgbxgh7CPkEctJHZMeM3rAvoW8CZKxtpG`
  */
 export async function getMinerPeerId (minerId, { maxAttempts = 5 } = {}) {
-  // Make a concurrent request to both sources: RPC node and smart contract
-  const [rpcNodeResult, contractResult] = await Promise.allSettled([
-    getMinerPeerIdFromRpcNode(minerId, { maxAttempts }),
+  // Make a concurrent request to both sources: FilecoinMinerInfo and smart contract
+  const [minerInfoResult, contractResult] = await Promise.allSettled([
+    getMinerPeerIdFromFilecoinMinerInfo(minerId, { maxAttempts }),
     getMinerPeerIdFromSmartContract(minerId)
   ])
   // Check contract result first
@@ -40,13 +40,13 @@ export async function getMinerPeerId (minerId, { maxAttempts = 5 } = {}) {
     return contractResult.value
   }
 
-  // Fall back to rpc-Node result
-  if (rpcNodeResult.status === 'fulfilled' && rpcNodeResult.value) {
-    return rpcNodeResult.value
+  // Fall back to FilecoinMinerInfo result
+  if (minerInfoResult.status === 'fulfilled' && minerInfoResult.value) {
+    return minerInfoResult.value
   }
 
   // Handle the case where both failed
-  throw new Error(`Failed to resolve PeerID from any source: ${contractResult.reason}, ${rpcNodeResult.reason}`)
+  throw new Error(`Failed to resolve PeerID from any source: ${contractResult.reason}, ${minerInfoResult.reason}`)
 }
 
 /**
@@ -55,7 +55,7 @@ export async function getMinerPeerId (minerId, { maxAttempts = 5 } = {}) {
  * @param {number} [options.maxAttempts]
  * @returns {Promise<string>} Miner's PeerId, e.g. `12D3KooWMsPmAA65yHAHgbxgh7CPkEctJHZMeM3rAvoW8CZKxtpG`
  */
-export async function getMinerPeerIdFromRpcNode (minerId, { maxAttempts = 5 }) {
+export async function getMinerPeerIdFromFilecoinMinerInfo (minerId, { maxAttempts = 5 }) {
   const chainHead = await getChainHead({ maxAttempts })
   try {
     const res = await retry(() => rpc('Filecoin.StateMinerInfo', minerId, chainHead), {
