+++ lib/smart-contract-client.js
@@ -3,6 +3,8 @@ import { assert } from 'zinnia:assert'
 import { RPC_URL } from './constants.js'
 
 // ABI for the MinerPeerIDMapping contract (minimal ABI with just the method we need)
+// Docs for smart contract: https://github.com/filecoin-project/curio/blob/395bc47d0f585cbc869fd4671dc05b1b2f4b18c2/market/ipni/spark/sol/README.md
+// Reasoning for smart contract: https://docs.curiostorage.org/curio-market/ipni-interplanetary-network-indexer-provider#ipni-provider-identification
 const contractABI = [
   {
     inputs: [
@@ -39,8 +41,6 @@ const contractABI = [
 
 // Contract address on the Filecoin EVM
 const CONTRACT_ADDRESS = '0x14183aD016Ddc83D638425D6328009aa390339Ce'
-// Singleton instance of the smart contract client
-let smartContract
 // Initialize provider and contract
 async function getSmartContractClient (rpcUrl) {
   const provider = new ethers.providers.JsonRpcProvider(rpcUrl)
@@ -51,25 +51,19 @@ async function getSmartContractClient (rpcUrl) {
  * Query the smart contract for the peer ID mapping
  * @param {string} minerID - The miner ID (as string, will be converted to uint64)
  * @param {object} [options]
- * @param {function} [options.getSmartContractClientFn] - Function to get the smart contract client
+ * @param {function} [options.smartContract] - Function to get the smart contract client
  * @returns {Promise<string>} The peer ID from the contract or empty string if not found
  */
 export async function getMinerPeerIdFromSmartContract (
   minerID,
-  { getSmartContractClientFn } = {}
+  { smartContract } = {}
 ) {
   try {
-    // Reuse the smart contract if already initialized
-    // Use a custom initialization function if passed.
-    if (!smartContract || getSmartContractClientFn) {
-      smartContract = await (
-        getSmartContractClientFn ?? getSmartContractClient
-      )(RPC_URL)
-    }
-    assert(smartContract, 'smartContract must be initialized')
+    const contractClient = smartContract ?? (await getSmartContractClient(RPC_URL))
+    assert(contractClient, 'smartContract must be initialized')
     // Convert minerID string (like 'f01234') to numeric ID
     const numericID = parseInt(minerID.replace('f0', ''))
-    const peerData = await smartContract.getPeerData(numericID)
+    const peerData = await contractClient.getPeerData(numericID)
     // TODO: Check if peerData.signature is valid
     return peerData?.peerID ?? null
   } catch (error) {
