+++ test/spark.js
@@ -2,10 +2,8 @@
 
 import Spark, { calculateDelayBeforeNextTask, newStats } from '../lib/spark.js'
 import { test } from 'zinnia:test'
-import { assertInstanceOf, assertEquals, assertArrayIncludes, assertRejects, assert } from 'zinnia:assert'
-import { SPARK_VERSION, RPC_URL } from '../lib/constants.js'
-import { getMinerPeerIdFromSmartContract } from '../lib/smart-contract-client.js'
-import { getMinerPeerId } from '../lib/miner-info.js'
+import { assertInstanceOf, assertEquals, assertArrayIncludes } from 'zinnia:assert'
+import { SPARK_VERSION } from '../lib/constants.js'
 
 const KNOWN_CID = 'bafkreih25dih6ug3xtj73vswccw423b56ilrwmnos4cbwhrceudopdp5sq'
 
@@ -364,103 +362,4 @@ test('calculateDelayBeforeNextTask() handles one task per round', () => {
     lastTaskDurationInMs: 1_000
   })
   assertEquals(delay, 60_000)
-})
-
-const mockPeerIdResponse = {
-  peerID: '12D3KooWGQmdpbssrYHWFTwwbKmKL3i54EJC9j7RRNb47U9jUv1U',
-  signature: '0x1234567890abcdef'
-}
-
-const mockEmptyPeerIdResponse = {
-  peerID: '',
-  signature: '0x'
-}
-
-// Mock contract factory
-function createMockContract (mockResponses) {
-  return {
-    getPeerData: async (minerId) => {
-      const response = mockResponses[minerId]
-      if (!response) {
-        throw new Error(`Miner ID ${minerId} not found in contract`)
-      }
-      return response
-    }
-  }
-}
-
-test('smart-contract-client: getMinerPeerIdFromSmartContract returns peer ID for valid miner ID', async () => {
-  // Create mock contract with predefined responses
-  const peerId = 12345
-  const mockContract = createMockContract({
-    [peerId]: mockPeerIdResponse
-  })
-
-  const actualPeerId = await getMinerPeerIdFromSmartContract(`f0${peerId}`, {
-    getSmartContractClientFn: () => mockContract
-  })
-
-  assertEquals(actualPeerId, mockPeerIdResponse.peerID)
-})
-
-test('miner-info: getMinerPeerId integration test with real miner f01234', async () => {
-  const peerId = await getMinerPeerIdFromSmartContract('f012345', RPC_URL)
-  assertEquals(typeof peerId, 'string', 'Expected peerId to be a string')
-  assert(peerId.length > 0, 'Expected peerId to be non-empty')
-  assertEquals(peerId, '12D3KooWGQmdpbssrYHWFTwwbKmKL3i54EJC9j7RRNb47U9jUv1U')
-})
-
-test('smart-contract-client: getMinerPeerIdFromSmartContract returns empty string for miner ID with no peer ID', async () => {
-  // Create mock contract with predefined responses
-  const peerId = 99999
-  const mockContract = createMockContract({
-    [peerId]: mockEmptyPeerIdResponse
-  })
-
-  const actualPeerId = await getMinerPeerIdFromSmartContract(`f0${peerId}`, {
-    getSmartContractClientFn: () => mockContract
-  })
-
-  assertEquals(actualPeerId, '')
-})
-
-test('smart-contract-client: getMinerPeerIdFromSmartContract throws error for non-existent miner ID', async () => {
-  // Create mock contract with predefined responses (empty to cause error)
-  const mockContract = createMockContract({})
-
-  await assertRejects(
-    async () => {
-      await getMinerPeerIdFromSmartContract('f055555', {
-        getSmartContractClientFn: () => mockContract
-      })
-    },
-    Error,
-    'Error fetching peer ID from contract for miner f055555'
-  )
-})
-
-test('smart-contract-client: getMinerPeerIdFromSmartContract properly strips f0 prefix', async () => {
-  // Create a mock that validates the minerId was correctly converted
-  let receivedMinerId = null
-
-  const mockContract = {
-    getPeerData: async (minerId) => {
-      receivedMinerId = minerId
-      return mockPeerIdResponse
-    }
-  }
-
-  await getMinerPeerIdFromSmartContract('f0123456', {
-    getSmartContractClientFn: () => mockContract
-  })
-
-  assertEquals(receivedMinerId, 123456)
-})
-
-test('miner-info: getMinerPeerId integration test with real miner f03303347', async () => {
-  const peerId = await getMinerPeerId('f03303347')
-
-  assert(typeof peerId === 'string', 'Expected peerId to be a string')
-  assert(peerId.length > 0, 'Expected peerId to be non-empty')
-  assertEquals(peerId, '12D3KooWGQmdpbssrYHWFTwwbKmKL3i54EJC9j7RRNb47U9jUv1U')
 })
